Google maps integration

+++ app/src/main/java/com/example/razomua/data/local/entity/ProfileEntity.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "profiles")
 data class ProfileEntity(
-    @PrimaryKey val id: Long,
-    val userId: Long,
+    @PrimaryKey val id: Int,
+    val userId: Int,
     val photoUrl: String?,
     val location: String?
 )
