+++ app/src/main/java/com/example/razomua/data/local/entity/SwipeEntity.kt
@@ -5,9 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "swipes")
 data class SwipeEntity(
-    @PrimaryKey val id: Long,
-    val fromUserId: Long,
-    val toUserId: Long,
+    @PrimaryKey val id: Int,
+    val fromUserId: Int,
+    val toUserId: Int,
     val action: String,
     val createdAt: String
 )
